Unit test de vertex
